fix: layout component
