+++ apps/web/package.json
@@ -19,6 +19,7 @@
     "@payloadcms/next-payload": "^0.1.11",
     "@payloadcms/richtext-slate": "^1.3.1",
     "drizzle-orm": "^0.29.1",
+    "lodash": "^4.17.21",
     "next": "^13.5.4",
     "payload": "^2.4.0",
     "react": "^18.2.0",
