+++ apps/web/src/app/(app)/layout.tsx
@@ -1,17 +1,23 @@
-import React, { PropsWithChildren } from "react";
+import React, { PropsWithChildren } from 'react';
 
-import ThemeProvider from "theme/src/ThemeProvider";
+import ThemeProvider from 'theme/src/ThemeProvider';
 
-import Layout from "@web/components/Layout";
-import StyledComponentsRegistry from "@web/lib/StyledComponentRegistry";
+import Layout from '@web/components/Layout';
+import fetchPayloadData from '@web/lib/fetchPayloadData';
+import StyledComponentsRegistry from '@web/lib/StyledComponentRegistry';
 
-export default function AppLayout({ children }: PropsWithChildren<object>) {
+export default async function AppLayout({
+  children
+}: PropsWithChildren<object>) {
+  const navigation = await fetchPayloadData((client) =>
+    client.findGlobal({ slug: 'navigation' })
+  );
   return (
     <html lang="en">
       <body>
         <StyledComponentsRegistry>
           <ThemeProvider>
-            <Layout>{children}</Layout>
+            <Layout navigation={navigation}>{children}</Layout>
           </ThemeProvider>
         </StyledComponentsRegistry>
       </body>
