+++ apps/web/src/components/Layout/__tests__/index.test.tsx
@@ -3,6 +3,7 @@ import { composeStories } from '@storybook/testing-react';
 import { render } from '@testing-library/react';
 
 import projectAnnotations from 'ui/utils/testSetup';
+
 import * as stories from '../Layout.stories';
 
 const { Defaults } = composeStories(stories, projectAnnotations);
