+++ apps/web/src/components/index.tsx
@@ -1,6 +1,8 @@
 import dynamic from 'next/dynamic';
 
-export const ImageWithContent = dynamic(() => import('./ImageWithContent')) as React.FC;
+export const ImageWithContent = dynamic(
+  () => import('./ImageWithContent')
+) as React.FC;
 
 // CODE INJECTION COMMENT
 export const Hero = dynamic(() => import('./Hero')) as React.FC;
