+++ packages/theme/src/theme.ts
@@ -1,9 +1,9 @@
 import { theme } from '@refract-ui/sc';
 
 const defaultTheme = theme({
-    themeColors: {
-        primary: '#FAFF00'
-    }
+  themeColors: {
+    primary: '#FAFF00'
+  }
 });
 
 export type RefractTheme = typeof defaultTheme;
